DOM Assignment update 1.0

+++ script.js
@@ -11,7 +11,7 @@ window.addEventListener("DOMContentLoaded", async () => {
             liElement.innerHTML = `Book Title: ${b.title}  Author: ${b.author}  ISBN: ${b.isbn} Availability: <input type="radio" class="avail" name="avail" value="true"/><label>Yes</label><input type="radio" class="avail" name="avail" value="false"/><label>No</label><button class="editBook" id="editBook" style="margin: 5px;">Edit</button><button class="deleteBook" id="deleteBook" style="margin: 5px;">Delete</button>`;
             display.append(liElement);
 
-            document.querySelector("#editBook").addEventListener("click", () => {
+           /* document.querySelector("#editBook").addEventListener("click", () => {
                 let isbn = prompt("Enter the ISBN of the book you want to edit");
                 let newTitle = prompt("Enter the title of book");
                 let newAuthor = prompt("Enter the name of author");
@@ -21,11 +21,16 @@ window.addEventListener("DOMContentLoaded", async () => {
             })
         
             document.querySelector("#deleteBook").addEventListener("click", () => {
-                let isbn = prompt("Enter the ISBN of the book to delete");
+                let toDelete = confirm("Do you really want to delete a book?");
+
+                if(toDelete) {
+                    let isbn = prompt("Enter the ISBN of the book to delete");
         
-                deleteBook(books, isbn);
-                renderList();
-            })
+                    deleteBook(books, isbn);
+                    renderList();
+                }
+
+            }) */
         
 
             let radioButton = liElement.querySelector(".avail");
@@ -67,10 +72,12 @@ window.addEventListener("DOMContentLoaded", async () => {
     // })
 
     // document.querySelector("#deleteBook").addEventListener("click", () => {
+    //      let toDelete = confirm("Do you really want to delete a book?");
+    //      if (toDelete) {
     //     let isbn = prompt("Enter the ISBN of the book to delete");
-
     //     deleteBook(books, isbn);
     //     renderList();
+            }     
     // })
 
     // document.querySelector("#save-btn").addEventListener("click", (books) => {
